Split the Pegasus provider into multiple files
Also decreased memory usage by removing statics.
